% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.add}
\alias{checkins.add}
\title{Create a check-in}
\usage{
checkins.add(venueId = NULL, eventId = NULL, shout = NULL,
  mentions = NULL, broadcast = NULL, ll = NULL, llAcc = NULL,
  alt = NULL, altAcc = NULL, m = "swarm")
}
\arguments{
\item{venueId}{required The venue where the user is checking
in. Find venue IDs by searching or from historical APIs.}

\item{eventId}{The event the user is checking in to.}

\item{shout}{A message about your check-in. The maximum
length of this field is 140 characters.}

\item{mentions}{Mentions in your check-in. This parameter is
a semicolon-delimited list of mentions. A single mention is
of the form "start,end,userid", where start is the index of
the first character in the shout representing the mention,
end is the index of the first character in the shout after
the mention, and userid is the userid of the user being
mentioned.  If userid is prefixed with "fbu-", this
indicates a Facebook userid that is being mention.
Character indices in shouts are 0-based.}

\item{broadcast}{Who to broadcast this check-in to. Accepts
a comma-delimited list of values:  private (off the grid)
or public (share with friends) facebook share on facebook
twitter share on twitter followers share with followers
(celebrity mode users only)   If no valid value is found,
the default is public.}

\item{ll}{Latitude and longitude of the user's location.
Only specify this field if you have a GPS or other device
reported location for the user at the time of check-in.}

\item{llAcc}{Accuracy of the user's latitude and longitude,
in meters.}

\item{alt}{Altitude of the user's location, in meters.}

\item{altAcc}{Vertical accuracy of the user's location, in
meters.}

\item{m}{Accepts values of "swarm"}
}
\value{
checkin A checkin object. A post-checkin
	notifications object.
}
\description{
Allows you to check in to a place.
}
\details{
Allows you to check in to a place.
	Note: This function requires acting user
}
\examples{
checkins.add(venueId="IHR8THISVNU", eventId="UHR8THISVNT",
	shout="There are crayons! Crayons!",
	mentions="5,10,HZXXY3Y;15,20,GZYYZ3Z;25,30,fbu-GZXY13Y",
	broadcast="public,twitter", ll="33.7,44.2", llAcc="1",
	alt="0", altAcc="1")
}

