% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.addpost}
\alias{checkins.addpost}
\title{Add a Post}
\usage{
checkins.addpost(CHECKIN_ID = NULL, text = NULL, url = NULL,
  contentId = NULL, m = "swarm")
}
\arguments{
\item{CHECKIN_ID}{The ID of the checkin to add a post to.}

\item{text}{The text of the post, up to 200 characters.}

\item{url}{Link for more details. This page will be opened
in an embedded web view in the foursquare application,
unless contentId is specified and a native link handler is
registered and present. We support the following URL
schemes: http, https, foursquare, mailto, tel, and sms.}

\item{contentId}{Identifier for the post to be used in a
native link, up to 50 characters. A url must also be
specified in the request.}

\item{m}{Accepts values of "swarm"}
}
\value{
post The newly-created post.
}
\description{
Post user generated content from an external app.
}
\details{
Post user generated content from an external app to a
	check-in. This post will be accessible to anyone who can
	view the details of the check-in.  Only call this endpoint
	if the user has chosen to post some information from the
	application. If you want to create a post containing a
	photo, use photos/add instead.
	Note: This function requires acting user
}
\examples{
checkins.addpost(CHECKIN_ID="IHR8THISVNU", text="Awesome!",
	url="http://your.site.com/path", contentId="3ER4GN93AFWE")
}

