% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.checkins}
\alias{checkins.checkins}
\title{Check-in Details}
\usage{
checkins.checkins(CHECKIN_ID = NULL, signature = NULL, m = "swarm")
}
\arguments{
\item{CHECKIN_ID}{The ID of the checkin to retrieve
additional information for.}

\item{signature}{When checkins are sent to public feeds such
as Twitter, foursquare appends a signature (s=XXXXXX)
allowing users to bypass the friends-only access check on
checkins. The same value can be used here for programmatic
access to otherwise inaccessible checkins. Callers should
use the bit.ly API to first expand 4sq.com links.}

\item{m}{Accepts values of "swarm"}
}
\value{
checkin A complete checkin object.
}
\description{
Get details of a checkin.
}
\details{
Get details of a checkin.
	Note: This function requires acting user
}
\examples{
checkins.checkins(CHECKIN_ID="IHR8THISVNU",
	signature="ASDJKASLJDLA")
}

