% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.deletecomment}
\alias{checkins.deletecomment}
\title{Remove commment from check-in}
\usage{
checkins.deletecomment(CHECKIN_ID = NULL, commentId = NULL, m = "swarm")
}
\arguments{
\item{CHECKIN_ID}{The ID of the checkin to remove a comment
from.}

\item{commentId}{The id of the comment to remove.}

\item{m}{Accepts values of "swarm"}
}
\value{
checkin The checkin, minus this comment.
}
\description{
Remove a comment from a checkin.
}
\details{
Remove a comment from a checkin, if the acting user is the
	author or the owner of the checkin.
	Note: This function requires acting user
}
\examples{
checkins.deletecomment(CHECKIN_ID="IHR8THISVNU",
	commentId="ABZQ200")
}

