% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.likes}
\alias{checkins.likes}
\title{Users who have liked a checkin}
\usage{
checkins.likes(CHECKIN_ID = NULL, m = "swarm")
}
\arguments{
\item{CHECKIN_ID}{The ID of the checkin to get likes for.}

\item{m}{Accepts values of "swarm"}
}
\value{
likes A count and groups of users who like this
	checkin. Groups generally include friends and others
	(indicating the relationship to the acting user), but are
	subject to change.
}
\description{
Returns friends and a total count of users who have liked.
}
\details{
Returns friends and a total count of users who have liked
	this checkin.
}
\examples{
checkins.likes(CHECKIN_ID="HZXXY3Y")
}

