% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins.recent}
\alias{checkins.recent}
\title{Recent checkins by friends}
\usage{
checkins.recent(ll = NULL, limit = NULL, afterTimestamp = NULL,
  m = "swarm")
}
\arguments{
\item{ll}{Latitude and longitude of the user's location, so
response can include distance.}

\item{limit}{Number of results to return, up to 100.}

\item{afterTimestamp}{Seconds after which to look for
checkins, e.g. for looking for new checkins since the last
fetch. If more than limit results are new since then, this
is ignored. Checkins created prior to this timestamp will
still be returned if they have new comments or photos,
making it easier to poll for all new activity.}

\item{m}{Accepts values of "swarm"}
}
\value{
checkins An array of checkin objects with user
	details present. Usually, venue details are also present,
	depending on whether the checkin has a venue.
}
\description{
Returns a list of recent checkins from friends.
}
\details{
Returns a list of recent checkins from friends.
	Note: This function requires acting user
}
\examples{
checkins.recent(ll="44.3,37.2", limit="20",
	afterTimestamp="123456")
}

