% Generated by roxygen2 (4.0.2): do not edit by hand
\name{events.add}
\alias{events.add}
\title{Add an event}
\usage{
events.add(venueId = NULL, name = NULL, startAt = NULL, endAt = NULL,
  m = "foursquare")
}
\arguments{
\item{venueId}{The id of the venue where the event is being
held.}

\item{name}{The name of the event.}

\item{startAt}{Time when the event is scheduled to start, in
seconds since Unix epoch.}

\item{endAt}{Time when the event is scheduled to end, in
seconds since Unix epoch.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
event An event object.
}
\description{
Create an event for a venue that you manage.
}
\details{
Create an event for a venue that you manage. You can see all
	your events in the tools tab when you're managing your
	venue.  They're on the calendar.  Events show up when users
	view your venue on mobile or the web. Users can also check
	in to events and share with their friends what they're up
	to in the present moment, and look back to see what they
	did in the days of yore.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
events.add(venueId="f8ea99941e", name=""Trivia Night!"",
	startAt="1355469600", endAt="1355470000")
}

