% Generated by roxygen2 (4.0.2): do not edit by hand
\name{events.events}
\alias{events.events}
\title{Event details}
\usage{
events.events(EVENT_ID = NULL, m = "foursquare")
}
\arguments{
\item{EVENT_ID}{required The ID of the event to retrieve
additional information for.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
event A complete event object.
}
\description{
Get details of a event.
}
\details{
Get details of a event.
	Note: This function requires acting user
}
\examples{
events.events(EVENT_ID="4d0fb8162d39a340637dc42b")
}

