% Generated by roxygen2 (4.0.2): do not edit by hand
\name{events.search}
\alias{events.search}
\title{Search Events}
\usage{
events.search(domain = NULL, eventId = NULL, participantId = NULL,
  m = "foursquare")
}
\arguments{
\item{domain}{required Identifier for a known third-party
event provider. This is used in conjunction with id.
Currently songkick.com is the only supported value.}

\item{m}{Accepts values of "foursquare" or "swarm"}

\item{eventId_or_participantId}{required Identifier used by
third-party specifed in domain, which we will attempt to
match against our events listings. eventId is the id of an
event, participantId is the id of a participant, like a
band or sports team.}
}
\value{
event An array of compact events.
}
\description{
Experimental
}
\details{
This is an experimental API and subject to change or
	breakage.  Returns a list of events matching the search
	parameters.
}
\examples{
events.search(domain="songkick.com",
	eventId_or_participantId="8183976")
}

