% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initFoursquare}
\alias{initFoursquare}
\title{Initialize foursquare OAuth}
\usage{
initFoursquare(CLIENT_ID, CLIENT_SECRET, ACCESS_TOKEN)
}
\arguments{
\item{CLIENT_ID}{User application client id}

\item{CLIENT_SECRET}{User application client secret}

\item{ACCESS_TOKEN}{Foursquare OAuth authorization token}
}
\description{
Initialize using user app credentials and access token
}

