% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.add}
\alias{lists.add}
\title{Add a List}
\usage{
lists.add(name = NULL, description = NULL, collaborative = NULL,
  photoId = NULL, m = "foursquare")
}
\arguments{
\item{name}{required The name of the list.}

\item{description}{The description of the list.}

\item{collaborative}{Boolean indicating if this list can be
edited by friends.}

\item{photoId}{The id of a photo that should be set as the
list photo.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list The list that was just created
}
\description{
Allows users to create a new list.
}
\details{
Allows users to create a new list.
	Note: This function requires acting user
}
\examples{
lists.add(name="Max's Favorite Ramen Spots",
	description="Must visit ramen spots according to Max.",
	collaborative="true", photoId="IHR8THISPHT")
}

