% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.additem}
\alias{lists.additem}
\title{Add an Item}
\usage{
lists.additem(LIST_ID = NULL, venueId = NULL, url = NULL, tipId = NULL,
  listId = NULL, itemId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created or followed
list as well as one of either USER_ID/tips or
USER_ID/todos.}

\item{venueId}{optional A venue to add to the list.}

\item{url}{optional If adding a new tip via text, this can
associate a url with the tip.}

\item{tipId}{optional Used to add a tip to a list. Cannot be
used in conjunction with the text and url fields.}

\item{listId}{optional Used in conjuction with itemId, the
id for a user created or followed list as well as one of
either USER_ID/tips or USER_ID/todos.}

\item{itemId}{optional Used in conjuction with listId, the
id of an item on that list that we wish to copy to this
list.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list item The list item that was just added.
}
\description{
Allows you to add an item to a list.
}
\details{
Allows you to add an item to a list.  All fields are
	optional, but exactly one of the following must be
	specified:  venueId tipId listId and itemId  Note: This
	endpoint is now the preferred way to add an item to a
	user's To-do list (/v2/lists/USER_ID/todos/additem).
	Note: This function requires acting user
}
\examples{
lists.additem(LIST_ID="HZXXY3Y", venueId="HZXXY3Y",
	url="http://nyt.com", tipId="HZXXY3Y", listId="HZXXY3Y",
	itemId="HZXXY3Y")
}

