% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.deleteitem}
\alias{lists.deleteitem}
\title{Delete an Item}
\usage{
lists.deleteitem(LIST_ID = NULL, itemId = NULL, venueId = NULL,
  tipId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created or followed
list or one of either USER_ID/tips or USER_ID/todos.}

\item{itemId}{optional id of the item to delete.}

\item{venueId}{optional id of a venue to be deleted. If the
venue is on the list multiple times, e.g. multiple tips at
the same venue, all items will be removed.}

\item{tipId}{optional id of a tip to be deleted.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list items A count and items of list item that were
	just deleted.
}
\description{
Allows you to delete items from a list.
}
\details{
Allows you to delete items from a list. One of itemId,
	venueId, or tipId must be provided.  Note: Collaborators
	can only delete items they added. List owners may delete
	any item.
	Note: This function requires acting user
}
\examples{
lists.deleteitem(LIST_ID="HZXXY3Y", itemId="HZXXY3Y",
	venueId="HZXXY3Y", tipId="HZXXY3Y")
}

