% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.follow}
\alias{lists.follow}
\title{Follow a List}
\usage{
lists.follow(LIST_ID = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id of a user-created list.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list The list that was just followed.
}
\description{
Allows you to follow a list.
}
\details{
Allows you to follow a list.  Note: Only valid on
	user-created lists.
	Note: This function requires acting user
}
\examples{
lists.follow(LIST_ID="HZXXY3Y")
}

