% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.followers}
\alias{lists.followers}
\title{List Followers}
\usage{
lists.followers(LIST_ID = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list}

\item{m}{Accepts values of "foursquare"}
}
\value{
followers Returns a pageable list of compact user.
}
\description{
Returns a count and items of users following this list.
}
\details{
Returns a count and items of users following this list.
	Note: Only valid on user-created lists
}
\examples{
lists.followers(LIST_ID="HZXXY3Y")
}

