% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.items}
\alias{lists.items}
\title{List Item Detail}
\usage{
lists.items(LIST_ID = NULL, ITEM_ID = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created or followed
list.}

\item{ITEM_ID}{required id for an item in the parent list.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list An item object.
}
\description{
Gives details about a list item.
}
\details{
Gives details about a list item
	Note: This function requires acting user
}
\examples{
lists.items(LIST_ID="A9ABCD", ITEM_ID="U7EU8R")
}

