% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.lists}
\alias{lists.lists}
\title{List Detail}
\usage{
lists.lists(USER_ID = NULL, defaultList = "todos", LIST_ID = NULL,
  limit = NULL, offset = NULL, llBounds = NULL, categoryId = NULL,
  sort = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created (e.g.
/v2/lists/12381902N) or followed list or one of either
USER_ID/tips (e.g. /v2/lists/32/tips) or USER_ID/todos.}

\item{limit}{Number of results to return, up to 200.}

\item{offset}{The number of results to skip. Used to page
through results.}

\item{llBounds}{optional Restricts the returned results to
the input bounding box.}

\item{categoryId}{optional Restricts the returned results to
venues matching the input category id.}

\item{sort}{optional Sorts the list items. Possible values
are recent and nearby. recent sorts the list items by the
date added to the list. nearby sorts the list items by the
distance from the center of the provided llBounds.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list A list object.
}
\description{
Gives details about a list.
}
\details{
Gives details about a list.  The list id parameter can be a
	user-created list id as well as one of either USER_ID/tips
	or USER_ID/todos. Note that a user's todos are only visible
	to their friends.
}
\examples{
lists.lists(LIST_ID="A9ABCD", limit="100", offset="100",
	llBounds="30.13800,-98.16009,30.40485,-97.28118",
	categoryId="4bf58dd8d48988d10e941735", sort="recent")
}

