% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.moveitem}
\alias{lists.moveitem}
\title{Move Item}
\usage{
lists.moveitem(LIST_ID = NULL, itemId = NULL, beforeId = NULL,
  afterId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id of a user-created list.}

\item{itemId}{required id of the item on this list to move.}

\item{beforeId}{optional Move itemId before beforeId.}

\item{afterId}{optional Move itemId after afterId.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list The list that was just edited.
}
\description{
Allows you to move an item on a list.
}
\details{
Allows you to move an item on a list. One of beforeId or
	afterId must be specified.  Note: Only valid on
	user-created lists
	Note: This function requires acting user
}
\examples{
lists.moveitem(LIST_ID="HZXXY3Y", itemId="HZXXY3Y",
	beforeId="HZXXY3Y", afterId="HZXXY3Y")
}

