% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.suggestphoto}
\alias{lists.suggestphoto}
\title{Suggest Photo}
\usage{
lists.suggestphoto(LIST_ID = NULL, itemId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list.}

\item{itemId}{required id of item on this list.}

\item{m}{Accepts values of "foursquare"}
}
\value{
photos Returns groups user and others containing
	lists (a count and items of photos) of photos uploaded by
	this user and uploaded by other users.
}
\description{
Suggests photos that may be appropriate for this item.
}
\details{
Suggests photos that may be appropriate for this item.
	Note: Only valid on user-created lists
	Note: This function requires acting user
}
\examples{
lists.suggestphoto(LIST_ID="HZXXY3Y", itemId="HZXXY3Y")
}

