% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.suggesttip}
\alias{lists.suggesttip}
\title{Suggest Tip}
\usage{
lists.suggesttip(LIST_ID = NULL, itemId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list}

\item{itemId}{required id of item}

\item{m}{Accepts values of "foursquare"}
}
\value{
tips Returns groups user and others containing lists
	(a count and items of tips) of tips created by this user
	and created by other users.
}
\description{
Suggests tips that may be appropriate for this item.
}
\details{
Suggests tips that may be appropriate for this item.  Note:
	Only valid on user-created lists
	Note: This function requires acting user
}
\examples{
lists.suggesttip(LIST_ID="HZXXY3Y", itemId="HZXXY3Y")
}

