% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.suggestvenues}
\alias{lists.suggestvenues}
\title{Suggest Venues}
\usage{
lists.suggestvenues(LIST_ID = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list}

\item{m}{Accepts values of "foursquare"}
}
\value{
suggestedVenues An array of compact venues that may
	be appropriate for this list.
}
\description{
Suggests venues that may be appropriate for this list.
}
\details{
Suggests venues that may be appropriate for this list.
	Note: Only valid on user-created lists
	Note: This function requires acting user
}
\examples{
lists.suggestvenues(LIST_ID="HZXXY3Y")
}

