% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.unfollow}
\alias{lists.unfollow}
\title{Unfollow a List}
\usage{
lists.unfollow(LIST_ID = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id of a user-created list.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list The list that was just unfollowed.
}
\description{
Allows you to unfollow a list.
}
\details{
Allows you to unfollow a list.  Note: Only valid on
	user-created lists.
	Note: This function requires acting user
}
\examples{
lists.unfollow(LIST_ID="HZXXY3Y")
}

