% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.update}
\alias{lists.update}
\title{Updates a list}
\usage{
lists.update(LIST_ID = NULL, name = NULL, description = NULL,
  collaborative = NULL, photoId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list.}

\item{name}{If present and a non-empty value, updates the
List name.}

\item{description}{If present and a non-empty value, updates
the List description. If present and empty, will remove the
List description.}

\item{collaborative}{Boolean indicating if this list can be
edited by friends. Once this has been set to true for a
list, authenticated friends can edit the  list via additem,
deleteitem, etc.}

\item{photoId}{If present and a non-empty value, updates the
List photo. If present and empty, will remove the List
photo.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list The list that was just edited.
}
\description{
Allows you to update a list.
}
\details{
Allows you to update a list.  Note: Only valid on
	user-created lists
	Note: This function requires acting user
}
\examples{
lists.update(LIST_ID="HZXXY3Y", name="Max's Favorite Ramen
	Spots", description="Must visit ramen spots according to
	Max.", collaborative="true", photoId="HZXXY3Y")
}

