% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lists.updateitem}
\alias{lists.updateitem}
\title{Update a List Item}
\usage{
lists.updateitem(LIST_ID = NULL, itemId = NULL, tipId = NULL,
  text = NULL, url = NULL, photoId = NULL, m = "foursquare")
}
\arguments{
\item{LIST_ID}{required id for a user-created list.}

\item{itemId}{required The id of an item on this list.}

\item{tipId}{optional If present and a non-empty value, adds
or replaces a tip on this item. If present and empty, will
remove the tip on this item.}

\item{text}{optional If present creates a public tip on the
venue and replaces any existing tip on the item. Cannot be
used in conjuction with tipId or photoId}

\item{url}{optional If adding a new tip via text, this can
associate a url with the tip.}

\item{photoId}{optional If present and a non-empty value,
adds a photo to this item. If present and empty, will
remove the photo on this item. If the photo was a private
checkin photo, it will be promoted to a public venue photo.}

\item{m}{Accepts values of "foursquare"}
}
\value{
list item The updated list item.
}
\description{
Allows you to add or remove photos and tips.
}
\details{
Allows you to add or remove photos and tips from items on
	user-created lists.  Note: Only valid on user-created lists
	 Note: Collaborators can only update items they added. List
	owners can not update any item.
	Note: This function requires acting user
}
\examples{
lists.updateitem(LIST_ID="HZXXY3Y", itemId="HZXXY3Y",
	tipId="HZXXY3Y", text="This ramen spot is the jam",
	url="http://nyt.com", photoId="HZXXY3Y")
}

