% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pages.add}
\alias{pages.add}
\title{Add a Page}
\usage{
pages.add(name = NULL, m = "foursquare")
}
\arguments{
\item{name}{required the name of the page}

\item{m}{Accepts values of "foursquare"}
}
\value{
user The user account for the page that was just
	created. An OAuth access token for this page for the acting
	OAuth consumer.
}
\description{
Allows users to create a new page.
}
\details{
Allows users to create a new page.  The creating user is
	added as a manager of the new page.
	Note: This function requires acting user
}
\examples{
pages.add(name="Habana Outpost")
}

