% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pageupdates.add}
\alias{pageupdates.add}
\title{Add a Page Update}
\usage{
pageupdates.add(pageId = NULL, groupId = NULL, venueId = NULL,
  shout = NULL, photoId = NULL, broadcast = NULL, m = "foursquare")
}
\arguments{
\item{pageId}{required id of the page to associate with the
broadcast.  To find the page for a venue you can look at
the page object in the venue response.}

\item{groupId}{The venue group from which to broadcast an
update.}

\item{venueId}{A venue from which to broadcast an update.}

\item{shout}{Text associated with the broadcast. 160
characters max, 10 characters min.}

\item{photoId}{An optional photo to attach to the broadcast.
 For a new photo, you should use the photo add endpoint and
specify only the pageId parameter.}

\item{broadcast}{Additional places to send the broadcast to.
Accepts a comma-delimited list of values:  facebook share
on facebook twitter share on twitter private just create
the update without broadcasting to anyone}

\item{m}{Accepts values of "foursquare"}
}
\value{
pageUpdate A page update object.
}
\description{
Broadcast an update as a page to followers of the page.
}
\details{
Broadcast an update as a page to followers of the page and
	associated venues. Venues can be specified either by the
	venueId, groupId or pageId (meaning all venues managed by
	the page) parameters. Broadcasts will show up on venue
	pages in our apps and website.  A broadcast must contain a
	shout and may contain a photo. The API enforces a rate
	limit of at most one broadcast sent by a given user to the
	same set of venues in 15 minutes. This rate may change and
	is stricter for larger sets of venues.
	Note: This function requires acting user
}
\examples{
pageupdates.add(pageId="23456", groupId="f8ea99941e",
	venueId="f8ea99941e", shout=""Try our new summer dish!"",
	photoId="f8ea99941e", broadcast="twitter")
}

