% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pageupdates.pageupdates}
\alias{pageupdates.pageupdates}
\title{Follow or unfollow a page.}
\usage{
pageupdates.pageupdates(UPDATE_ID = NULL, limit = NULL, ll = NULL,
  m = "foursquare")
}
\arguments{
\item{UPDATE_ID}{The ID of the update to retrieve additional
information for.}

\item{limit}{The number of nearby venues to show (default
20).}

\item{ll}{The location of the current user.}

\item{m}{Accepts values of "foursquare"}

\item{USER_ID}{required The page to follow or unfollow.}

\item{set}{If 1, follow this page. If 0 unfollow (un-do a
previous follow) it. Default value is 1.}

\item{m}{Accepts values of "foursquare"}
}
\value{
user A user object representing the page the acting
	user has just followed or unfollowed.

pageUpdate A page update object.
}
\description{
Allows the acting user to follow or unfollow a page.
}
\details{
Allows the acting user to follow or unfollow a page.
	Following a page subscribes the acting user to updates from
	that page.
	Note: This function requires acting user
}
\examples{
pages.follow(USER_ID="XXX123YYYY", set="1")
Page Update Details

 Get page update details.

Get page update details.
	Note: This function requires acting user
}

