% Generated by roxygen2 (4.0.2): do not edit by hand
\name{photos.photos}
\alias{photos.photos}
\title{Photo details}
\usage{
photos.photos(PHOTO_ID = NULL, m = "foursquare")
}
\arguments{
\item{PHOTO_ID}{required The ID of the photo to retrieve
additional information for.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
photo A complete photo object.
}
\description{
Get details of a photo.
}
\details{
Get details of a photo.
	Note: This function requires acting user
}
\examples{
photos.photos(PHOTO_ID="4d0fb8162d39a340637dc42b")
}

