% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings.set}
\alias{settings.set}
\title{Change a setting}
\usage{
settings.set(SETTING_ID = NULL, value = NULL, m = "foursquare")
}
\arguments{
\item{SETTING_ID}{required Name of setting to change,
sendMayorshipsToTwitter, sendBadgesToTwitter,
sendMayorshipsToFacebook, sendBadgesToFacebook,
receivePings, receiveCommentPings.}

\item{value}{required 1 for true, and 0 for false.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
message A confirmation message.
}
\description{
Change a setting for the given user.
}
\details{
Change a setting for the given user.
	Note: This function requires acting user
}
\examples{
settings.set(SETTING_ID="sendBadgesToTwitter", value="1")
}

