% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings.settings}
\alias{settings.settings}
\title{Add a Photo}
\usage{
settings.settings(SETTING_ID = NULL, m = "foursquare")
}
\arguments{
\item{SETTING_ID}{The name of a setting.}

\item{m}{Accepts values of "foursquare" or "swarm"}

\item{checkinId}{the ID of a checkin owned by the user.}

\item{tipId}{the ID of a tip owned by the user.}

\item{venueId}{the ID of a venue, provided only when adding
a public photo of the venue in general, rather than a photo
for a private checkin, tip, or page update.}

\item{pageId}{the ID of a page, provided only when adding a
photo that will be in an update for that page (no other ids
should be specified).}

\item{broadcast}{Whether to broadcast this photo. Send
twitter if you want to send to twitter, facebook if you
want to send to facebook, or twitter,facebook if you want
to send to both.}

\item{public}{When the checkinId is also provided (meaning
this is a photo attached to a checkin), this parameter
allows for making the photo public and viewable at the
venue. Valid values are 1 and 0 (default).Note that photos
attached to venues, tips, and pages (updates) are always
public.}

\item{ll}{Latitude and longitude of the user's location.}

\item{llAcc}{Accuracy of the user's latitude and longitude,
in meters.}

\item{alt}{Altitude of the user's location, in meters.}

\item{altAcc}{Vertical accuracy of the user's location, in
meters.}

\item{postUrl}{A link for more details about the photo. This
page will be opened in an embedded web view in the
foursquare application, unless contentId is specified and a
native link handler is registered and present. We support
the following URL schemes: http, https, foursquare, mailto,
tel, and sms.}

\item{postContentId}{Identifier for the photo post to be
used in a native link, up to 50 characters. A checkinId and
postUrl must also be specified in the request.}

\item{postText}{Text for the photo post, up to 200
characters. A checkinId must also be specified in the
request.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
photo The photo that was just created.

value The value for this setting for the acting
	user.
}
\description{
Allows users to add a new photo to a checkin, tip, venue, etc.
}
\details{
Allows users to add a new photo to a checkin, tip, venue, or
	page update in general.  All fields are optional, but
	exactly one of the id fields (checkinId, tipId, venueId,
	pageId) must be passed in.  In addition, the image file
	data must be posted. The photo should be uploaded as a jpeg
	and the Content-Type should be set to "image/jpeg". Images
	may not exceed 5MB in size.  Attaching a photo to a tip or
	a venue makes it visible to anybody. By default, attaching
	a photo to a checkin makes it visible only to the people
	who can see the checkin (the user's friends, unless the
	checkin has been sent to Twitter or Facebook); this can be
	overridden by specifying public=1 in the request.
	Attaching a photo to a page is visible to anybody and is
	intended for use only in an update.  Multiple photos can be
	attached to a checkin, venue, or page, but there can only
	be one photo per tip.  To avoid double-tweeting, if you are
	sending a checkin that will be immediately followed by a
	photo, do not set broadcast=twitter on the checkin, and
	just set it on the photo.  When adding a photo to a
	checkin, specify the post* params to connect the photo back
	to the corresponding content inside your app. For more
	information, see our documentation on native links.
	Note: This function requires acting user
}
\examples{
photos.add(checkinId="IHR8THISVNU", tipId="IHR8THISVNU",
	venueId="IHR8THISVNU", pageId="2345", broadcast="twitter",
	public="1", ll="33.7,44.2", llAcc="1", alt="0", altAcc="1",
	postUrl="http://your.site.com/path",
	postContentId="3ER4GN93AFWE", postText="Awesome!")
Setting detail

 Returns a setting for the acting user.

Returns a setting for the acting user.
	Note: This function requires acting user
}

