% Generated by roxygen2 (4.0.2): do not edit by hand
\name{specials.add}
\alias{specials.add}
\title{Create a new special}
\usage{
specials.add(name = NULL, text = NULL, finePrint = NULL, count1 = NULL,
  type = NULL, offerId = NULL, cost = NULL, m = "foursquare")
}
\arguments{
\item{name}{A name for the special.}

\item{text}{Required. Maximum length of 200 characters.}

\item{finePrint}{Maximum length of 200 characters. Fine
print, shown in small type on the special detail page.}

\item{count1}{Specifier for special types.}

\item{type}{Required. The type of special.
frequencyCheck-in Special if count1 is 1; otherwise a
Loyalty Specialunlocked every count1 check-ins countNewbie
Special if count1 is 1; otherwise a Loyalty Specialunlocked
on the count1th check-in (all-time)}

\item{offerId}{Maximum length of 16 characters. Internal id
in your 3rd party system.}

\item{cost}{The amount of money the user must spend to use
this special in dollars and cents. For example, 5.50
meaning 5 dollars and 50 cents.}

\item{m}{Accepts values of "foursquare"}
}
\value{
special A special object.
}
\description{
Allows you to create a new special.
}
\details{
Allows you to create a new special. As of November 2013,
	mayor, regular, swarm, friends, and flash specials are
	deprecated.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
specials.add(name="Mayor Special", text="Congrats, as the
	Mayor, you get a free side!", finePrint="Does not include
	fries.", count1="5", type="frequency", offerId="5000000",
	cost="5.50")
}

