% Generated by roxygen2 (4.0.2): do not edit by hand
\name{specials.search}
\alias{specials.search}
\title{Search Specials}
\usage{
specials.search(ll = NULL, radius = NULL, llAcc = NULL, alt = NULL,
  altAcc = NULL, limit = NULL, m = "foursquare")
}
\arguments{
\item{ll}{Required. Latitude and longitude to search near.}

\item{radius}{Limit results to venues within this many
meters of the specified location. Defaults to a city-wide
area.}

\item{llAcc}{Accuracy of latitude and longitude, in meters.}

\item{alt}{Altitude of the user's location, in meters.}

\item{altAcc}{Accuracy of the user's altitude, in meters.}

\item{limit}{Number of results to return, up to 50.}

\item{m}{Accepts values of "foursquare"}
}
\value{
specials An array of specials being run at nearby
	venues. This array is identical in structure to the nearby
	specials array returned by the venue detail endpoint. Each
	special in the returned array contains a compact venue.
}
\description{
Returns a list of specials near the current location.
}
\details{
Returns a list of specials near the current location.
}
\examples{
specials.search(ll="44.3,37.2", radius="800",
	llAcc="10000.0", alt="0", altAcc="10000.0", limit="10")
}

