% Generated by roxygen2 (4.0.2): do not edit by hand
\name{specials.specials}
\alias{specials.specials}
\title{Special Detail}
\usage{
specials.specials(SPECIAL_ID = NULL, venueId = NULL, userId = NULL,
  m = "foursquare")
}
\arguments{
\item{SPECIAL_ID}{required ID of special to retrieve}

\item{venueId}{required ID of a venue the special is running
at}

\item{userId}{ID of the user to check whether the special is
unlocked for. Only available if the current user is the
manager of the venue. If not provided, checks wher the
special is unlocked for the current user.}

\item{m}{Accepts values of "foursquare"}
}
\value{
special A complete special.
}
\description{
Gives details about a special.
}
\details{
Gives details about a special, including text and whether it
	is unlocked for the current or provided user.
}
\examples{
specials.specials(SPECIAL_ID="A9ABCD", venueId="B92CD1",
	userId="2462")
}

