% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.add}
\alias{tips.add}
\title{Add a tip}
\usage{
tips.add(venueId = NULL, text = NULL, url = NULL, broadcast = NULL,
  m = "foursquare")
}
\arguments{
\item{venueId}{required The venue where you want to add this
tip.}

\item{text}{required The text of the tip, up to 200
characters.}

\item{url}{A URL related to this tip.}

\item{broadcast}{Whether to broadcast this tip. Send twitter
if you want to send to twitter, facebook if you want to
send to facebook, or twitter,facebook if you want to send
to both.}

\item{m}{Accepts values of "foursquare"}
}
\value{
tip The newly-added tip.
}
\description{
Allows you to add a new tip at a venue.
}
\details{
Allows you to add a new tip at a venue.
	Note: This function requires acting user
}
\examples{
tips.add(venueId="XXX123YYYY", text="The donuts are tasty!",
	url="http://blog.zagat.com/fast-food-survey-results-are-her
	e", broadcast="twitter")
}

