% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.like}
\alias{tips.like}
\title{Like or unlike a tip}
\usage{
tips.like(TIP_ID = NULL, set = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{required The tip to like or unlike.}

\item{set}{If 1, like this tip. If 0 unlike (un-do a
previous like) it. Default value is 1.}

\item{m}{Accepts values of "foursquare"}
}
\value{
likes Updated count and groups of users who like
	this tip. Groups generally include friends and others
	(indicating the relationship to the acting user), but are
	subject to change.
}
\description{
Allows the acting user to like or unlike a tip.
}
\details{
Allows the acting user to like or unlike a tip.
	Note: This function requires acting user
}
\examples{
tips.like(TIP_ID="XXX123YYYY", set="1")
}

