% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.likes}
\alias{tips.likes}
\title{Users who have liked a tip}
\usage{
tips.likes(TIP_ID = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{The ID of the tip to get likes for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
likes A count and groups of users who like this tip.
	Groups generally include friends and others (indicating the
	relationship to the acting user), but are subject to
	change.
}
\description{
Returns friends and a total count of users who have liked this tip
}
\details{
Returns friends and a total count of users who have liked
	this tip.
}
\examples{
tips.likes(TIP_ID="HZXXY3Y")
}

