% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.listed}
\alias{tips.listed}
\title{Lists}
\usage{
tips.listed(TIP_ID = NULL, group = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{Identity of a tip to get lists for.}

\item{group}{can be created, edited, followed, friends,
other. If no acting user is present, only other is
supported.}

\item{m}{Accepts values of "foursquare"}
}
\value{
lists If group is specified, it will contain count
	and items of lists. If no group is specified, it will
	contain a groups array containing elements, each with type,
	name, an optional count, and an optional items.  The type
	field can be passed in as a group to get more elements. If
	present count represents the total number of elements.  If
	not specified, there are a potentially unbounded number of
	elements. items contains a sample of the lists.
}
\description{
The lists that this tip appears on.
}
\details{
The lists that this tip appears on
}
\examples{
tips.listed(TIP_ID="HZXXY3Y", group="edited")
}

