% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.saves}
\alias{tips.saves}
\title{Users who have saved a tip}
\usage{
tips.saves(TIP_ID = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{The ID of the tip to get saves for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
saves A count and groups of users who saved this
	tip. Groups generally include friends and others
	(indicating the relationship to the acting user), but are
	subject to change.
}
\description{
Returns friends and a total count of users who have saved this tip.
}
\details{
Returns friends and a total count of users who have saved
	this tip.
}
\examples{
tips.saves(TIP_ID="HZXXY3Y")
}

