% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.tips}
\alias{tips.tips}
\title{Tip detail}
\usage{
tips.tips(TIP_ID = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{required ID of tip to retrieve}

\item{m}{Accepts values of "foursquare"}
}
\value{
tip A complete tip.
}
\description{
Gives details about a tip.
}
\details{
Gives details about a tip, including which users (especially
	friends) have marked the tip to-do.
}
\examples{
tips.tips(TIP_ID="A9ABCD")
}

