% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tips.unmark}
\alias{tips.unmark}
\title{Unmark a tip as to-do}
\usage{
tips.unmark(TIP_ID = NULL, m = "foursquare")
}
\arguments{
\item{TIP_ID}{required The tip you want to unmark.}

\item{m}{Accepts values of "foursquare"}
}
\value{
tip The tip being acted on.
}
\description{
Allows you to remove a tip from your to-do list.
}
\details{
Allows you to remove a tip from your to-do list.  Note: The
	lists deleteitem endpoint is now the preferred way to
	remove a tip from a user's To-do list
	(/v2/lists/USER_ID/todos/deleteitem).
	Note: This function requires acting user
}
\examples{
tips.unmark(TIP_ID="XXX123YYYY")
}

