% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updates.updates}
\alias{updates.updates}
\title{Update detail}
\usage{
updates.updates(UPDATE_ID = NULL, m = "foursquare")
}
\arguments{
\item{UPDATE_ID}{required The ID of the update to retrieve}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
notification A update.
}
\description{
Update detail
}
\details{
Update detail
	Note: This function requires acting user
}
\examples{
updates.updates(UPDATE_ID="BTEP567")
}

