% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.approve}
\alias{users.approve}
\title{Approve a friend request}
\usage{
users.approve(USER_ID = "self", m = "swarm")
}
\arguments{
\item{USER_ID}{required The user ID of a pending friend.}

\item{m}{Accepts values of "swarm"}
}
\value{
user A user object for the approved user.
}
\description{
Approves a pending friend request from another user.
}
\details{
Approves a pending friend request from another user.
	Note: This function requires acting user
}
\examples{
users.approve(USER_ID="HZXXY3Y")
}

