% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.checkins}
\alias{users.checkins}
\title{Checkins by a user}
\usage{
users.checkins(USER_ID = "self", limit = NULL, offset = NULL,
  sort = NULL, afterTimestamp = NULL, beforeTimestamp = NULL,
  m = "swarm")
}
\arguments{
\item{USER_ID}{For now, only self is supported}

\item{limit}{Number of results to return, up to 250.}

\item{offset}{The number of results to skip. Used to page
through results.}

\item{sort}{How to sort the returned checkins. Can be
newestfirst or oldestfirst.}

\item{afterTimestamp}{Retrieve the first results to follow
these seconds since epoch. This should be useful for paging
forward in time, or when polling for changes. To avoid
missing results when polling, we recommend subtracting
several seconds from the last poll time and then
de-duplicating.}

\item{beforeTimestamp}{Retrieve the first results prior to
these seconds since epoch. Useful for paging backward in
time.}

\item{m}{Accepts values of "swarm"}
}
\value{
checkins A count and items of check-ins.
}
\description{
Returns a history of checkins for the authenticated user.
}
\details{
Returns a history of checkins for the authenticated user.
	Note: We highly discourage developers from polling this
	endpoint to learn of new check-ins or to keep their users'
	check-in history synced. Using the real-time API is a much
	better solution.
	Note: This function requires acting user
}
\examples{
users.checkins(USER_ID="self", limit="100", offset="100",
	sort="newestfirst", afterTimestamp="1279044824",
	beforeTimestamp="1279044824")
}

