% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.deny}
\alias{users.deny}
\title{Deny a friend request}
\usage{
users.deny(USER_ID = "self", m = "swarm")
}
\arguments{
\item{USER_ID}{required The user ID of a pending friend.}

\item{m}{Accepts values of "swarm"}
}
\value{
user A user object for the denied user.
}
\description{
Denies a pending friend request from another user.
}
\details{
Denies a pending friend request from another user.
	Note: This function requires acting user
}
\examples{
users.deny(USER_ID="HZXXY3Y")
}

