% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.friends}
\alias{users.friends}
\title{List friends}
\usage{
users.friends(USER_ID = "self", limit = NULL, offset = NULL,
  m = "swarm")
}
\arguments{
\item{USER_ID}{Identity of the user to get friends of. Pass
self to get friends of the acting user.}

\item{limit}{Number of results to return, up to 500.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "swarm"}
}
\value{
friends A count and items of compact user objects.
}
\description{
Returns an array of a user's friends.
}
\details{
Returns an array of a user's friends.
	Note: This function requires acting user
}
\examples{
users.friends(USER_ID="HZXXY3Y", limit="100", offset="100")
}

