% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.lists}
\alias{users.lists}
\title{Lists}
\usage{
users.lists(USER_ID = "self", group = NULL, ll = NULL, limit = NULL,
  offset = NULL, m = "foursquare")
}
\arguments{
\item{USER_ID}{Identity of the user to get lists for. Pass
self to get lists of the acting user.}

\item{group}{can be created (lists created by this user),
edited (other people's lists this user has edited),
followed (lists this user follows), friends (lists from
this user's friends), and suggested (lists relevant to the
user's current location).}

\item{ll}{Location of the user, required in order to receive
the suggested group.}

\item{limit}{Number of results to return, up to 200.}

\item{offset}{The number of results to skip. Used to page
through results.}

\item{m}{Accepts values of "foursquare"}
}
\value{
lists If group is specified, contains a count and
	items of lists. If no group is specified, it contains a
	groups array containing elements, each with type, name, an
	optional count and optional items. The type field can be
	passed in as a group to get more elements. If present,
	count represents the total number of elements. If not
	specified, there are a potentially unbounded number of
	elements. items contains a sample of the items.
}
\description{
A User's Lists.
}
\details{
A User's Lists.
	Note: This function requires acting user
}
\examples{
users.lists(USER_ID="HZXXY3Y", group="edited",
	ll="40.7,-74", limit="100", offset="100")
}

