% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.mayorships}
\alias{users.mayorships}
\title{List mayorships}
\usage{
users.mayorships(USER_ID = "self", m = "swarm")
}
\arguments{
\item{USER_ID}{Identity of the user to get mayorships for.
Pass self to get friends of the acting user.}

\item{m}{Accepts values of "swarm"}
}
\value{
mayorships A count and items of objects which
	currently only contain compact venue objects.
}
\description{
Returns a user's mayorships.
}
\details{
Returns a user's mayorships.
	Note: This function requires acting user
}
\examples{
users.mayorships(USER_ID="HZXXY3Y")
}

