% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.photos}
\alias{users.photos}
\title{Photos from a User}
\usage{
users.photos(USER_ID = "self", limit = NULL, offset = NULL,
  m = "foursquare")
}
\arguments{
\item{USER_ID}{For now, only self is supported.}

\item{limit}{Number of results to return, up to 500.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
photos A count and items of photos.
}
\description{
Returns photos from a user.
}
\details{
Returns photos from a user.
	Note: This function requires acting user
}
\examples{
users.photos(USER_ID="self", limit="100", offset="100")
}

