% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.search}
\alias{users.search}
\title{Find users}
\usage{
users.search(phone = NULL, email = NULL, twitter = NULL,
  twitterSource = NULL, fbid = NULL, name = NULL, m = "foursquare")
}
\arguments{
\item{phone}{A comma-delimited list of phone numbers to look
for.}

\item{email}{A comma-delimited list of email addresses to
look for.}

\item{twitter}{A comma-delimited list of Twitter handles to
look for.}

\item{twitterSource}{A single Twitter handle. Results will
be users that this handle follows on Twitter who use
Foursquare.}

\item{fbid}{A comma-delimited list of Facebook ID's to look
for.}

\item{name}{A single string to search for in users' names.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
results An array of compact user objects with
	Twitter or Facebook information and friend status. An array
	of unmatched search terms grouped into subarrays
	corresponding to the term's parameter.
}
\description{
Helps a user locate friends.
}
\details{
Helps a user locate friends.
	Note: This function requires acting user
}
\examples{
users.search(phone="91712344567,123450",
	email="foo@bar.com,alice@baz.com",
	twitter="dens,sesamestreet", twitterSource="krave",
	fbid="123,456", name="Fred Fredrickson")
}

