% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.setpings}
\alias{users.setpings}
\title{Set whether to receive pings about a user}
\usage{
users.setpings(USER_ID = "self", value = NULL, m = "swarm")
}
\arguments{
\item{USER_ID}{required The user ID of a friend.}

\item{value}{required True or false.}

\item{m}{Accepts values of "swarm"}
}
\value{
user A user object for the user.
}
\description{
Changes whether the acting user will receive pings.
}
\details{
Changes whether the acting user will receive pings (phone
	notifications) when the specified user checks in.
	Note: This function requires acting user
}
\examples{
users.setpings(USER_ID="HZXXY3", value="false")
}

