% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.tips}
\alias{users.tips}
\title{Tips from a User}
\usage{
users.tips(USER_ID = "self", sort = NULL, ll = NULL, limit = NULL,
  offset = NULL, m = "foursquare")
}
\arguments{
\item{USER_ID}{Identity of the user to get tips from. Pass
self to get tips of the acting user.}

\item{sort}{One of recent, nearby, or popular. Nearby
requires geolat and geolong to be provided.}

\item{ll}{Latitude and longitude of the user's location.}

\item{limit}{Number of results to return, up to 500.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "foursquare"}
}
\value{
tips A count and items of tips.
}
\description{
Returns tips from a user.  This endpoint is deprecated.
}
\details{
Returns tips from a user.  This endpoint is deprecated. Use
	/lists/USER_ID/tips instead.
	Note: This function requires acting user
}
\examples{
users.tips(USER_ID="HZXXY3Y", sort="recent", ll="33.7,44.2",
	limit="100", offset="100")
}

