% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.unfriend}
\alias{users.unfriend}
\title{Remove a Friend}
\usage{
users.unfriend(USER_ID = "self", m = "swarm")
}
\arguments{
\item{USER_ID}{Identity of the user to unfriend.}

\item{m}{Accepts values of "swarm"}
}
\value{
user A user.
}
\description{
Cancels relationship of the acting user.
}
\details{
Cancels any relationship between the acting user and the
	specified user.  Removes a friend, unfollows a celebrity,
	or cancels a pending friend request.
	Note: This function requires acting user
}
\examples{
users.unfriend(USER_ID="HZXXY3")
}

