% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.users}
\alias{users.users}
\title{User}
\usage{
users.users(USER_ID = "self", m = "foursquare")
}
\arguments{
\item{USER_ID}{Identity of the user to get details for. Pass
self to get details of the acting user.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
user A user.
}
\description{
Returns profile information for a given user.
}
\details{
Returns profile information for a given user, including
 selected badges and mayorships. The web profile for a user
	is visible at https://foursquare.com/user/USER_ID  If the
	user is a friend, contact information, Facebook ID, and
	Twitter handle and the user's last checkin may also be
	present.  In addition, the pings field will indicate
	whether checkins from this user will trigger a ping
	(notifications to mobile devices). This setting can be
	changed via setpings.  Note that this setting is overriden
	if pings is false in settings (no pings will be sent, even
	if this user is set to true).
	Note: This function requires acting user
}
\examples{
users.users(USER_ID="HZXXY3")
}

