% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.venuehistory}
\alias{users.venuehistory}
\title{Venues visited by a user}
\usage{
users.venuehistory(USER_ID = "self", beforeTimestamp = NULL,
  afterTimestamp = NULL, categoryId = NULL, m = "swarm")
}
\arguments{
\item{USER_ID}{For now, only self is supported}

\item{beforeTimestamp}{Seconds since epoch.}

\item{afterTimestamp}{Seconds after epoch.}

\item{categoryId}{Limits returned venues to those in this
category. If specifying a top-level category, all
sub-categories will also match the query.}

\item{m}{Accepts values of "swarm"}
}
\value{
venues A count and items of objects containing a
	beenHere count and venue compact venues.
}
\description{
Returns a list of all venues visited by the specified user.
}
\details{
Returns a list of all venues visited by the specified user,
	along with how many visits and when they were last there.
	Note: This function requires acting user
}
\examples{
users.venuehistory(USER_ID="self",
	beforeTimestamp="123456789", afterTimestamp="123456789",
	categoryId="asad13242ljla")
}

