% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users.venuelikes}
\alias{users.venuelikes}
\title{Venues liked by a user}
\usage{
users.venuelikes(USER_ID = "self", beforeTimestamp = NULL,
  afterTimestamp = NULL, categoryId = NULL, limit = NULL, offset = NULL,
  m = "foursquare")
}
\arguments{
\item{USER_ID}{User ID or self}

\item{beforeTimestamp}{Seconds since epoch.}

\item{afterTimestamp}{Seconds since epoch.}

\item{categoryId}{Limits returned venues to those in this
category. If specifying a top-level category, all
sub-categories will also match the query.}

\item{limit}{Number of results to return.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venues A count and items of objects containing a
	beenHere count and venue compact venues.
}
\description{
Returns a list of venues liked by the specified user.
}
\details{
Returns a list of venues liked by the specified user
	Note: This function requires acting user
}
\examples{
users.venuelikes(USER_ID="self",
	beforeTimestamp="123456789", afterTimestamp="123456789",
	categoryId="asad13242ljla", limit="100", offset="100")
}

