% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.add}
\alias{venuegroups.add}
\title{Create a new venue group}
\usage{
venuegroups.add(name = NULL, venueId = NULL, m = "foursquare")
}
\arguments{
\item{name}{Required. The name to give the group.}

\item{venueId}{Comma-delimited list of venue IDs to add to
the group. If this parameter is not specified, then the
venue group will initially be empty.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venue group A venue group object.
}
\description{
Create a venue group.
}
\details{
Create a venue group. If the venueId parameter is specified,
	then the endpoint will add the specified venues to the
	venue group. If it is not possible to add all of the
	specified venues to the group, then creation of the venue
	group will fail entirely.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.add(name="Think Coffees", venueId="IHR8THISVNU")
}

