% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.addvenue}
\alias{venuegroups.addvenue}
\title{Add Venue}
\usage{
venuegroups.addvenue(GROUP_ID = NULL, venueId = NULL, m = "foursquare")
}
\arguments{
\item{GROUP_ID}{required The ID of the venue group to modify}

\item{venueId}{required comma-delimited list of venue IDs to
add to the group}

\item{m}{Accepts values of "foursquare"}
}
\value{
Success code or error message
}
\description{
Add a venue to a venue group.
}
\details{
Add a venue to a venue group.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.addvenue(GROUP_ID="IHR8THISVNU",
	venueId="IHR8THISVNU")
}

