% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.delete}
\alias{venuegroups.delete}
\title{Delete a Venue Group}
\usage{
venuegroups.delete(GROUP_ID = NULL, m = "foursquare")
}
\arguments{
\item{GROUP_ID}{required The ID of the venue group to
delete.}

\item{m}{Accepts values of "foursquare"}
}
\value{
Success code or error message
}
\description{
Delete a venue group.
}
\details{
Delete a venue group.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.delete(GROUP_ID="IHR8THISVNU")
}

