% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.timeseries}
\alias{venuegroups.timeseries}
\title{Venue Group Time Series Data}
\usage{
venuegroups.timeseries(GROUP_ID = NULL, startAt = NULL, endAt = NULL,
  fields = NULL, m = "foursquare")
}
\arguments{
\item{GROUP_ID}{The venue group to retrieve series data for.}

\item{startAt}{required. The start of the time range to
retrieve stats for (seconds since epoch).}

\item{endAt}{The end of the time range to retrieve stats for
(seconds since epoch). If omitted, the current time is
assumed.}

\item{fields}{Specifies which fields to return. May be one
or more of totalCheckins, newCheckins, uniqueVisitors,
sharing, genders, ages, hours, separated by commas.}

\item{m}{Accepts values of "foursquare"}
}
\value{
timeseries An array of venue time series data
	objects, one for each venue in the group.
}
\description{
Get daily venue stats for the venues in a group.
}
\details{
Get daily venue stats for the venues in a group over a time
	range.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.timeseries(GROUP_ID="AGRP234",
	startAt="1284286794", endAt="1284286794",
	fields="totalCheckins,newCheckins")
}

