% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.update}
\alias{venuegroups.update}
\title{Update Venue Group}
\usage{
venuegroups.update(GROUP_ID = NULL, name = NULL, venueId = NULL,
  m = "foursquare")
}
\arguments{
\item{GROUP_ID}{required The ID of the venue group to modify}

\item{name}{If specified, the new name to give to the group.}

\item{venueId}{If specified, a comma-delimited list of venue
IDs that will become the new set of venue IDs for the
group.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venue group A venue group object.
}
\description{
Updates a venue group.
}
\details{
Updates a venue group. At least one of the name and venueId
	parameters must be specified.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.update(GROUP_ID="IHR8THISVNU", name="Think
	Coffees", venueId="IHR8THISVNU")
}

