% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venuegroups.venuegroups}
\alias{venuegroups.venuegroups}
\title{Venue Group Details}
\usage{
venuegroups.venuegroups(GROUP_ID = NULL, m = "foursquare")
}
\arguments{
\item{GROUP_ID}{The ID of the venue group to retrieve
additional information for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venuegroup A venue group object.
}
\description{
Get venue group details.
}
\details{
Get venue group details.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venuegroups.venuegroups(GROUP_ID="IHR8THISVNU")
}

