% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.add}
\alias{venues.add}
\title{Add a Venue}
\usage{
venues.add(name = NULL, address = NULL, crossStreet = NULL, city = NULL,
  state = NULL, zip = NULL, phone = NULL, twitter = NULL, ll = NULL,
  primaryCategoryId = NULL, description = NULL, url = NULL,
  ignoreDuplicates = NULL, ignoreDuplicatesKey = NULL, m = "foursquare")
}
\arguments{
\item{name}{required the name of the venue}

\item{address}{The address of the venue.}

\item{crossStreet}{The nearest intersecting street or
streets.}

\item{city}{The city name where this venue is.}

\item{state}{The nearest state or province to the venue.}

\item{zip}{The zip or postal code for the venue.}

\item{phone}{The phone number of the venue.}

\item{twitter}{The twitter handle of the venue.}

\item{ll}{required Latitude and longitude of the venue, as
accurate as is known.}

\item{primaryCategoryId}{The ID of the category to which you
want to assign this venue.}

\item{description}{A freeform description of the venue, up
to 160 characters.}

\item{url}{The url of the homepage of the venue.}

\item{ignoreDuplicates}{A boolean flag telling the server to
ignore duplicates and force the addition of this venue.}

\item{ignoreDuplicatesKey}{Required if ignoreDuplicates is
true. This key will be available in the response of the
HTTP 409 error of the first (failed) attempt to add venue.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venue The venue that was just created.
}
\description{
Allows Foursquare users to add a new venue.
}
\details{
Allows Foursquare users to add a new venue.   All fields are
	optional, except for ll and name.  Category ID’s Although
	optional, we strongly recommend that developers pass in the
	primaryCategoryId parameter to assign the new venue a
	category. The resulting venues become more meaningful
	within Foursquare and are more easily searchable by other
	users. We recommend that applications show their users our
	category hierarchy (obtained from venues/categories) and
	allow them to choose something suitable.  Duplicate Venue
	Errors The method may return an HTTP 409 error if the new
	venue looks like a duplicate of an existing venue. This
	error response will include two useful values:
	candidateDuplicateVenues and ignoreDuplicatesKey.   In this
	situation, you can either:  Use one of the
	candidateDuplicateVenues included in the response of the
	409 error. This will not create a new venue. Ignore
	duplicates and force the addition of a new venue by
	resubmitting the same venue add request with two additional
	parameters: ignoreDuplicates set to true and
	ignoreDuplicatesKey set to the value from the earlier error
	response.   The decision to ignore duplicates should be up
	to your users—developers should not always just pass
	ignoreDuplicates=true.  Adding Venues Without
	Authenticating Users First This endpoint generally requires
	you to authenticate Foursquare users before you can add
	venues on their behalf. In some cases, we’ll make
	exceptions and allow applications to create new venues
	without authenticating any users. If you’re interested in
	this feature, please contact api@foursquare.com.
	Note: This function requires acting user
}
\examples{
venues.add(name="Habana Outpost", address="1313 Mockingbird
	Lane", crossStreet="at Fulton St", city="New York",
	state="New York", zip="AE1234", phone="00 01 23 1234",
	twitter="eathabana", ll="44.3,37.2",
	primaryCategoryId="4bf58dd8d48988d1d4941735",
	description="We are a family owned and operated business
	with our customers' satisfaction ... ",
	url="http://www.mercurylounge.com",
	ignoreDuplicates="true",
	ignoreDuplicatesKey="bb29f248166648fb51d8d92b1740444c")
}

