% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.claim}
\alias{venues.claim}
\title{Claim a Venue}
\usage{
venues.claim(visible = NULL, m = "foursquare")
}
\arguments{
\item{visible}{(optional, default true) whether role as
manager is visible on the venue page}

\item{m}{Accepts values of "foursquare"}
}
\value{
Empty response
}
\description{
Claim a venue for the user.
}
\details{
Claim a venue for the user. If your OAuth Consumer is set up
	with the proper privileges, you can use this endpoint to
	claim a venue on behalf of a user without the claim having
	to be approved by foursquare staff.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venues.claim(visible="false")
}

