% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.events}
\alias{venues.events}
\title{Venue Events}
\usage{
venues.events(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue id for which events are
being requested.}

\item{m}{Accepts values of "foursquare"}
}
\value{
events A count and items of event items. Also
	includes a "summary" string describing the set of events at
	the venue.
}
\description{
Allows you to access information about the current events.
}
\details{
Allows you to access information about the current events at
	a place.  At this time we are only able to distribute Music
	events and limited information about Movie events via this
	endpoint due to partner restrictions.
}
\examples{
venues.events(VENUE_ID="AVNU234")
}

