% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.herenow}
\alias{venues.herenow}
\title{Venue Here Now}
\usage{
venues.herenow(VENUE_ID = NULL, limit = NULL, offset = NULL,
  m = "swarm")
}
\arguments{
\item{VENUE_ID}{required ID of venue to retrieve}

\item{limit}{Number of results to return, up to 500.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "swarm"}
}
\value{
hereNow A count and items where items are checkins.
}
\description{
Provides a count of how many people are at a given venue.
}
\details{
Provides a count of how many people are at a given venue.
	Returns a list of  friends and friends-of-friends at the
	venue for authenticated requests. If the acting user is
	currently checked in to the venue, then a list of other
	users is also returned.  Authenticated requests by the
	manager of a venue will always return the list of users
	“here now” for that venue  Do not aggregate information
	from this endpoint across multiple venues.  This endpoint
	is part of the venues API.
	Note: This function requires acting user
}
\examples{
venues.herenow(VENUE_ID="A9ABCD", limit="100", offset="100")
}

