% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.hours}
\alias{venues.hours}
\title{Venue Hours}
\usage{
venues.hours(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue id for which hours are
being requested.}

\item{m}{Accepts values of "foursquare"}
}
\value{
hours An array of timeframes of open hours. An array
	of timeframes of popular hours.
}
\description{
Returns hours for a venue.
}
\details{
Returns hours for a venue.
}
\examples{
venues.hours(VENUE_ID="XXX123YYYY")
}

