% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.like}
\alias{venues.like}
\title{Like or unlike a venue}
\usage{
venues.like(VENUE_ID = NULL, set = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue to like or unlike.}

\item{set}{If 1, like this venue. If 0 unlike (un-do a
previous like) it. Default value is 1.}

\item{m}{Accepts values of "foursquare"}
}
\value{
likes Updated count and groups of users who like
	this venue. Groups generally include friends and others
	(indicating the relationship to the acting user), but are
	subject to change.
}
\description{
Allows the acting user to like or unlike a venue.
}
\details{
Allows the acting user to like or unlike a venue.
	Note: This function requires acting user
}
\examples{
venues.like(VENUE_ID="XXX123YYYY", set="1")
}

