% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.likes}
\alias{venues.likes}
\title{Users who have liked a venue}
\usage{
venues.likes(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{The ID of the venue to get likes for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
likes A count and groups of users who like this
	venue. Groups generally include friends and others
	(indicating the relationship to the acting user), but are
	subject to change.
}
\description{
Returns friends and a total count of users who have liked.
}
\details{
Returns friends and a total count of users who have liked
	this venue.
}
\examples{
venues.likes(VENUE_ID="HZXXY3Y")
}

