% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.links}
\alias{venues.links}
\title{Links For a Venue}
\usage{
venues.links(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue you want annotations for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
links A count and items of links.
}
\description{
Returns URLs or identifiers from third parties.
}
\details{
Returns URLs or identifiers from third parties that have
	been applied to this venue, such as how the New York Times
	refers to this venue and a URL for additional information
	from nytimes.com. This is part of the foursquare Venue Map.
}
\examples{
venues.links(VENUE_ID="XXX123YYYY")
}

