% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.listed}
\alias{venues.listed}
\title{Lists}
\usage{
venues.listed(VENUE_ID = NULL, group = NULL, limit = NULL,
  offset = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{Identity of a venue to get lists for.}

\item{group}{can be created, edited, followed, friends,
other. If no acting user is present, only other is
supported.}

\item{limit}{Number of results to return, up to 200.}

\item{offset}{Used to page through results. Must specify a
group}

\item{m}{Accepts values of "foursquare"}
}
\value{
lists If group is specified, it will contain count
	and items of lists. If no group is specified, it will
	contain a groups array containing elements, each with type,
	name, an optional count, and an optional items.  The type
	field can be passed in as a group to get more elements. If
	present count represents the total number of elements.  If
	not specified, there are a potentially unbounded number of
	elements. items contains a sample of the lists.
}
\description{
The lists that this venue appears on.
}
\details{
The lists that this venue appears on
}
\examples{
venues.listed(VENUE_ID="HZXXY3Y", group="edited",
	limit="30", offset="300")
}

