% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.managed}
\alias{venues.managed}
\title{Managed Venues}
\usage{
venues.managed(limit = NULL, offset = NULL, m = "foursquare")
}
\arguments{
\item{limit}{Number of managed venues to return. Defaults to
100, maximum is 1000.}

\item{offset}{Number of venues to skip over for paging.
Defaults to 0.}

\item{m}{Accepts values of "foursquare"}
}
\value{
venues A count and items of compact venues the user
	manages.
}
\description{
Get a list of venues the current user manages.
}
\details{
Get a list of venues the current user manages.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venues.managed(limit="100", offset="500")
}

