% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.menu}
\alias{venues.menu}
\title{Venue Menu}
\usage{
venues.menu(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue id for which menu is
being requested.}

\item{m}{Accepts values of "foursquare"}
}
\value{
provider The name and attribution information for
	the menu provider. If attribution information is provided,
	you must display both the attributionImage and
	attributionText as well as link to attributionLink. A count
	and items of menu.
}
\description{
Returns menu information for a venue.
}
\details{
Returns menu information for a venue.   In some cases, menu
	information is provided by our partners. When displaying
	the information from a partner, you must attribute them
	using the attribution information included in the provider
	field. Not all menu information available on Foursquare is
	able to be redistributed through our API.
}
\examples{
venues.menu(VENUE_ID="AVNU234")
}

