% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.nextvenues}
\alias{venues.nextvenues}
\title{Next Venues}
\usage{
venues.nextvenues(VENUE_ID = NULL, m = "swarm")
}
\arguments{
\item{VENUE_ID}{required ID of the venue you want to see
next venue information about}

\item{m}{Accepts values of "swarm"}
}
\value{
nextVenues A count and items where items are compact
	venues.
}
\description{
Returns other venues that people often check.
}
\details{
Returns venues that people often check in to after the
	current venue. Up to 5 venues are returned in each query,
	and results are sorted by how many people have visited that
	venue after the current one. Homes are never returned in
	results.
}
\examples{
venues.nextvenues(VENUE_ID="A9ABCD")
}

