% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.photos}
\alias{venues.photos}
\title{Photos from a Venue}
\usage{
venues.photos(VENUE_ID = NULL, group = NULL, limit = NULL,
  offset = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue you want photos for.}

\item{group}{If not specified, public venue photos are
returned ordered by relevance. Pass venue for public venue
photos, ordered by recency. Pass checkin for venue photos
from friends (including non-public photos from recent
checkins), ordered by recency. See our documentation on
photos for information on how to handle the response and
construct actual photo URLs.}

\item{limit}{Number of results to return, up to 200.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "foursquare"}
}
\value{
photos A count and items of photo.
}
\description{
Returns photos for a venue.
}
\details{
Returns photos for a venue.
}
\examples{
venues.photos(VENUE_ID="XXX123YYYY", group="checkin",
	limit="100", offset="100")
}

