% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.search}
\alias{venues.search}
\title{Search Venues}
\usage{
venues.search(ll = NULL, near = NULL, llAcc = NULL, alt = NULL,
  altAcc = NULL, query = NULL, limit = NULL, intent = NULL,
  radius = NULL, sw = NULL, ne = NULL, categoryId = NULL, url = NULL,
  providerId = NULL, linkedId = NULL, m = "foursquare")
}
\arguments{
\item{ll}{required unless near is provided. Latitude and
longitude of the user's location. (Required for query
searches). Optional if using intent=global}

\item{near}{required unless ll is provided. A string naming
a place in the world. If the near string is not geocodable,
returns a failed_geocode error. Otherwise, searches within
the bounds of the geocode. Adds a geocode object to the
response. (Required for query searches)}

\item{llAcc}{Accuracy of latitude and longitude, in meters.
(Does not currently affect search results.)}

\item{alt}{Altitude of the user's location, in meters. (Does
not currently affect search results.)}

\item{altAcc}{Accuracy of the user's altitude, in meters.
(Does not currently affect search results.)}

\item{query}{A search term to be applied against venue
names.}

\item{limit}{Number of results to return, up to 50.}

\item{intent}{One of the values below, indicating your
intent in performing the search. If no value is specified,
defaults to checkin.   checkin Finds results that the
current user (or, for userless requests, a typical user) is
likely to check in to at the provided ll at the current
moment in time. This is the intent we recommend most apps
use.   browse Find venues within a given area. Unlike the
checkin intent, browse searches an entire region instead of
only finding Venues closest to a point. You must define a
region to search be including either the ll and radius
parameters, or the sw and ne. The region will be a
spherical cap if you include the ll and radius parameters,
or it will be a bounding quadrangle if you include the sw
and ne parameters.   global Finds the most globally
relevant venues for the search, independent of location.
Ignores all other parameters other than query and limit.
match Finds venues that are are nearly-exact matches for
the given parameters. This intent is highly sensitive to
the provided location. We recommend using this intent only
when trying to correlate an existing place database with
Foursquare's. The results will be sorted best match first,
taking distance and spelling mistakes/variations into
account. query and ll are the only required parameters for
this intent, but matching also supports phone, address,
city, state, zip, and twitter. There's no specified format
for these parameters—we do our best to normalize them and
drop them from the search if unsuccessful.}

\item{radius}{Limit results to venues within this many
meters of the specified location. Defaults to a city-wide
area. Only valid for requests with intent=browse, or
requests with intent=checkin and categoryId or query. Does
not apply to match intent requests. The maximum supported
radius is currently 100,000 meters.}

\item{sw}{With ne, limits results to the bounding quadrangle
defined by the latitude and longitude given by sw as its
south-west corner, and ne as its north-east corner. The
bounding quadrangle is only supported for intent=browse
searches. Not valid with ll or radius. Bounding quadrangles
with an area up to approximately 10,000 square kilometers
are supported.}

\item{ne}{See sw}

\item{categoryId}{A comma separated list of categories to
limit results to. If you specify categoryId specifying a
radius may improve results. If specifying a top-level
category, all sub-categories will also match the query.
Does not apply to match intent requests.}

\item{url}{A third-party URL which we will attempt to match
against our map of venues to URLs.}

\item{providerId}{Identifier for a known third party that is
part of our map of venues to URLs, used in conjunction with
linkedId.}

\item{linkedId}{1002207971611 Identifier used by third party
specified in providerId, which we will attempt to match
against our map of venues to URLs.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
venues An array of compact venues.
}
\description{
Returns a list of venues near the current location.
}
\details{
Returns a list of venues near the current location,
	optionally matching a search term.  To ensure the best
	possible results, pay attention to the intent parameter
	below. And if you're looking for "top" venues or
	recommended venues, use the explore endpoint instead.  If
	lat and long is provided, each venue includes a distance.
	If authenticated, the method will return venue metadata
	related to you and your friends. If you do not
	authenticate, you will not get this data.  Note that most
	of the fields returned inside venue can be optional. The
	user may create a venue that has no address, city or state
	(the venue is created instead at the geolat/geolong
	specified). Your client should handle these conditions
	safely.  You'll also notice a stats block that reveals some
	count data about the venue. herenow shows the number of
	people currently there (this value can be 0).  This
	endpoint is part of the venues API.
}
\examples{
venues.search(ll="44.3,37.2", near="Chicago, IL",
	llAcc="10000.0", alt="0", altAcc="10000.0", query="donuts",
	limit="10", intent="checkin", radius="800", sw="44.3,37.2",
	ne="44.1,37.4", categoryId="asad13242l,btbe24353m",
	url="http://nymag.com/listings/restaurant/abistro/",
	providerId="nymag", linkedId="4247")
}

