% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.similar}
\alias{venues.similar}
\title{Similar Venues}
\usage{
venues.similar(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue you want similar venues
for.}

\item{m}{Accepts values of "foursquare"}
}
\value{
similarVenues A count and items of similar venues.
}
\description{
Returns a list of venues similar to the specified venue.
}
\details{
Returns a list of venues similar to the specified venue.
	Note: This function requires acting user
}
\examples{
venues.similar(VENUE_ID="XXX123YYYY")
}

