% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.suggestcompletion}
\alias{venues.suggestcompletion}
\title{Suggest Completion Venues}
\usage{
venues.suggestcompletion(ll = NULL, near = NULL, llAcc = NULL,
  alt = NULL, altAcc = NULL, query = NULL, limit = NULL,
  radius = NULL, sw = NULL, ne = NULL, m = "foursquare")
}
\arguments{
\item{ll}{required Latitude and longitude of the user's
location. (Required for query searches)}

\item{near}{required unless ll is provided. A string naming
a place in the world. If the near string is not geocodable,
returns a failed_geocode error. Otherwise, searches within
the bounds of the geocode. Adds a geocode object to the
response. (Required for query searches)}

\item{llAcc}{Accuracy of latitude and longitude, in meters.
(Does not currently affect search results.)}

\item{alt}{Altitude of the user's location, in meters. (Does
not currently affect search results.)}

\item{altAcc}{Accuracy of the user's altitude, in meters.
(Does not currently affect search results.)}

\item{query}{required A search term to be applied against
titles. Must be at least 3 characters long.}

\item{limit}{Number of results to return, up to 100.}

\item{radius}{Limit results to venues within this many
meters of the specified location. Defaults to a city-wide
area. The maximum supported radius is currently
80,000 meters.}

\item{sw}{With ne, limits results to the bounding quadrangle
defined by the latitude and longitude given by sw as its
south-west corner, and ne as its north-east corner. The
bounding quadrangle is only supported for intent=browse
searches. Not valid with ll or radius. Bounding quadrangles
with an area up to approximately 10,000 square kilometers
are supported.}

\item{ne}{See sw}

\item{m}{Accepts values of "foursquare"}
}
\value{
minivenues An array of venues that only contain id,
	name, location, and categories.
}
\description{
Returns a list of mini-venues.
}
\details{
Returns a list of mini-venues partially matching the search
	term, near the location.
}
\examples{
venues.suggestcompletion(ll="44.3,37.2", near="Chicago, IL",
	llAcc="10000.0", alt="0", altAcc="10000.0",
	query="foursqu", limit="30", radius="800", sw="44.3,37.2",
	ne="44.1,37.4")
}

