% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.timeseries}
\alias{venues.timeseries}
\title{Venue Time Series Data}
\usage{
venues.timeseries(venueId = NULL, startAt = NULL, endAt = NULL,
  fields = NULL, m = "foursquare")
}
\arguments{
\item{venueId}{A comma-separated list of venue ids to
retrieve series data for. The current user must be the
manager of all venues specified.}

\item{startAt}{required. The start of the time range to
retrieve stats for (seconds since epoch).}

\item{endAt}{The end of the time range to retrieve stats for
(seconds since epoch). If omitted, the current time is
assumed.}

\item{fields}{Specifies which fields to return. May be one
or more of totalCheckins, newCheckins, uniqueVisitors,
sharing, genders, ages, hours, separated by commas.}

\item{m}{Accepts values of "foursquare"}
}
\value{
timeseries An array of venue time series data
	objects, one for each venue.
}
\description{
Get daily venue stats for a list of venues.
}
\details{
Get daily venue stats for a list of venues over a time
	range.
	Note: This function requires acting user
	User restrictions: Must be venue manager
}
\examples{
venues.timeseries(venueId="AVNU234,AVNU567",
	startAt="1284286794", endAt="1284286794",
	fields="totalCheckins,newCheckins")
}

