% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.tips}
\alias{venues.tips}
\title{Tips from a Venue}
\usage{
venues.tips(VENUE_ID = NULL, sort = NULL, limit = NULL, offset = NULL,
  m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required The venue you want tips for.}

\item{sort}{One of friends, recent, or popular.}

\item{limit}{Number of results to return, up to 500.}

\item{offset}{Used to page through results.}

\item{m}{Accepts values of "foursquare"}
}
\value{
tips A count and items of tips.
}
\description{
Returns tips for a venue.
}
\details{
Returns tips for a venue.
}
\examples{
venues.tips(VENUE_ID="XXX123YYYY", sort="recent",
	limit="100", offset="100")
}

