% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.trending}
\alias{venues.trending}
\title{Trending Venues}
\usage{
venues.trending(ll = NULL, limit = NULL, radius = NULL, m = "swarm")
}
\arguments{
\item{ll}{required Latitude and longitude of the user's
location.}

\item{limit}{Number of results to return, up to 50.}

\item{radius}{Radius in meters, up to approximately 2000
meters.}

\item{m}{Accepts values of "swarm"}
}
\value{
venues An array of venues that are currently
	trending, with their hereNow populated.
}
\description{
Returns a list of venues near the current location.
}
\details{
Returns a list of venues near the current location with the
	most people currently checked in.  This endpoint is part of
	the venues API.
}
\examples{
venues.trending(ll="44.3,37.2", limit="10", radius="100")
}

