% Generated by roxygen2 (4.0.2): do not edit by hand
\name{venues.venues}
\alias{venues.venues}
\title{Update user's photo}
\usage{
venues.venues(VENUE_ID = NULL, m = "foursquare")
}
\arguments{
\item{VENUE_ID}{required ID of venue to retrieve}

\item{m}{Accepts values of "swarm"}

\item{photo}{Photo under 100KB in multipart MIME encoding
with content type image/jpeg, image/gif, or image/png.}

\item{m}{Accepts values of "foursquare" or "swarm"}
}
\value{
user The current user object.

venue A complete venue.
}
\description{
Updates the user's profile photo.
}
\details{
Updates the user's profile photo.
	Note: This function requires acting user
}
\examples{
users.update(photo	="file.jpg")
Venue Detail

 Gives details about a venue.

Gives details about a venue, including location, mayorship,
	tags, tips, specials, and category.  Authenticated users
	will also receive information about who is here now.  If
	the venue ID given is one that has been merged into another
	"master" venue, the response will show data about the
	"master" instead of giving you an error.
}

